api/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
google-cloud-bigquery==3.20.1
pyarrow==15.0.2